+++ include/libtropic_l3.h
@@ -414,11 +414,13 @@ lt_ret_t lt_out__ecc_key_read(lt_handle_t *h, const lt_ecc_slot_t slot);
  * @param h           Device's handle
  * @param key         Buffer for retrieving a key; length depends on the type of key in the slot (32B for Ed25519, 64B
  * for P256), according to *curve*
+ * @param max_size    Size of the key buffer
  * @param curve       Will be filled by curve type
  * @param origin      Will be filled by origin type
  * @return            LT_OK if success, otherwise returns other error code.
  */
-lt_ret_t lt_in__ecc_key_read(lt_handle_t *h, uint8_t *key, lt_ecc_curve_type_t *curve, lt_ecc_key_origin_t *origin);
+lt_ret_t lt_in__ecc_key_read(lt_handle_t *h, uint8_t *key, const uint8_t max_size, lt_ecc_curve_type_t *curve,
+                             lt_ecc_key_origin_t *origin);
 
 /**
  * @brief Encodes ECC_Key_Erase command payload.