+++ tests/functional/lt_test_rev_ecc_key_generate.c
@@ -44,7 +44,7 @@ static lt_ret_t lt_test_rev_ecc_key_generate_cleanup(void)
         }
 
         LT_LOG_INFO("Reading slot #%" PRIu8 " (should fail)", i);
-        ret = lt_ecc_key_read(g_h, i, read_pub_key, &curve, &origin);
+        ret = lt_ecc_key_read(g_h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin);
         if (LT_L3_ECC_INVALID_KEY != ret) {
             LT_LOG_ERROR("Return value is not LT_L3_ECC_INVALID_KEY.");
             return ret;
@@ -99,7 +99,8 @@ void lt_test_rev_ecc_key_generate(lt_handle_t *h)
         LT_LOG_INFO("Testing ECC key slot #%" PRIu8 "...", i);
 
         LT_LOG_INFO("Checking if slot is empty...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Generating private key using P256 curve...");
         LT_TEST_ASSERT(LT_OK, lt_ecc_key_generate(h, i, TR01_CURVE_P256));
@@ -111,7 +112,7 @@ void lt_test_rev_ecc_key_generate(lt_handle_t *h)
         LT_TEST_ASSERT(LT_L3_FAIL, lt_ecc_key_generate(h, i, TR01_CURVE_ED25519));
 
         LT_LOG_INFO("Reading the generated public key...");
-        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Checking curve type of the read key...");
         LT_TEST_ASSERT(1, (curve == TR01_CURVE_P256));
@@ -130,7 +131,8 @@ void lt_test_rev_ecc_key_generate(lt_handle_t *h)
         LT_LOG_INFO("Testing ECC key slot #%" PRIu8 "...", i);
 
         LT_LOG_INFO("Checking if slot is empty...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Generating private key using Ed25519 curve...");
         LT_TEST_ASSERT(LT_OK, lt_ecc_key_generate(h, i, TR01_CURVE_ED25519));
@@ -142,7 +144,7 @@ void lt_test_rev_ecc_key_generate(lt_handle_t *h)
         LT_TEST_ASSERT(LT_L3_FAIL, lt_ecc_key_generate(h, i, TR01_CURVE_P256));
 
         LT_LOG_INFO("Reading the generated public key...");
-        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Checking curve type of the read key...");
         LT_TEST_ASSERT(1, (curve == TR01_CURVE_ED25519));
@@ -154,7 +156,8 @@ void lt_test_rev_ecc_key_generate(lt_handle_t *h)
         LT_TEST_ASSERT(LT_OK, lt_ecc_key_erase(h, i));
 
         LT_LOG_INFO("Trying to read the erased slot (should fail)...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
     }
     LT_LOG_LINE();