+++ CHANGELOG.md
@@ -124,6 +124,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `lt_get_info_fw_bank()`: Parameter `read_size` to indicate the number of read bytes from the FW bank.
 - `lt_get_log_req()`: Parameter `max_size` to check whether the output buffer is big enough.
 - `lt_r_mem_data_read()`: Parameter `max_size` to check whether the output buffer is big enough.
+- `lt_ecc_key_read()`: Parameter `max_size` to check whether the output buffer is big enough.
 
 ### Fixed
 - `lt_r_mem_data_write()`, `lt_out__r_mem_data_write()`: Mark `data` as `const`.